fix githube pages (404.html and redirect)
I managed to completely fix the behavior of forced refreshing of the user's page via the address bar. Since GitHube doesn't support routing on a multi-page site, we had to add some code by redirecting through the 404.html file to the main index.html of the application, and fix the private route “Favorites”. Now, when the user (if he is logged in) is forced to refresh the page, the user remains on the same page he was on (for example, on the Favorites page). If the user is not logged in, the standard routing behavior works - the user is redirected to the Home page.